Refactor into model.board and model.map
